Adjust logics for modules and some parts of a main window
